+++ app/assets/stylesheets/auth.scss
@@ -61,6 +61,7 @@
   display: flex;
   flex: 1;
   justify-content: center;
+  height: 100%;
 }
 
 .input-data {
