+++ app/assets/stylesheets/company.scss
@@ -6,7 +6,7 @@
 }
 
 .content-middle-body {
-  height: 80%;
+  height: 100vw;
   display: flex;
 }
