+++ app/assets/stylesheets/layout.scss
@@ -15,3 +15,7 @@
 	width: 18%;
 	padding: 5px;
 }
+
+.search-bar{
+	padding: 19px 0 69px 0;
+}
