+++ app/javascript/packs/application.js
@@ -5,6 +5,7 @@
 
 import 'bootstrap'
 import "@hotwired/turbo-rails"
+import "@fontsource/roboto";
 require("@rails/ujs").start()
 // prequire("@rails/activestorage").start()
 require("channels")
