+++ app/views/account/products/edit.html.erb
@@ -34,7 +34,7 @@
             </div>
 
             <div class="item">
-                            <%= form.file_field :images, :multiple => true, direct_upload: true %>
+              <%= form.file_field :images, :multiple => true, direct_upload: true %>
 
               <div class="image-upload">
