+++ app/views/account/products/new.html.erb
@@ -4,7 +4,7 @@
   <div class="wrapper">
     <div class="box-auth">
       <div class="box-wrapper">
-        <%= form_with model: [:advertise, @product], id: "form-register" , local: true do |form| %>
+        <%= form_with model: [:account, @product], id: "form-register" , local: true do |form| %>
           <div class="container stretch">
             <div class="form-register item">
               <div class="advertise-details">
