+++ app/views/advertise/products/_description_product.html.erb
@@ -1,0 +1,6 @@
+<h1 class="title-advertise">
+  <%= @product.title  %>
+</h1>
+<h2 class="price-advertise">
+  Valor: <%= number_to_currency(@product.price) %>
+</h2>
