+++ app/views/shared/_layout.html.erb
@@ -1,1 +1,5 @@
-<div class="navbar-top"></div>
+<div class="navbar-top">
+  <div class="search-bar">
+
+  </div>
+</div>
