+++ config/routes.rb
@@ -10,12 +10,14 @@ Rails.application.routes.draw do
   namespace :account do
     root 'home#index'
     resources :users, only: %i[edit update show destroy]
+    resources :products do
+      get 'desactive', on: :member
+      get 'active', on: :member
+    end
   end
 
   namespace :advertise do
     resources :products
-    get 'desactive/:id', to: 'products#desactive', as: 'product_desactive'
-    get 'active/:id', to: 'products#active', as: 'product_active'
   end
 
   # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
