+++ package.json
@@ -2,6 +2,7 @@
   "name": "marketplace",
   "private": true,
   "dependencies": {
+    "@fontsource/roboto": "^4.5.8",
     "@rails/actioncable": "^6.0.0",
     "@rails/activestorage": "^6.0.0",
     "@rails/ujs": "^6.0.0",
